==> _scripts/7_lms.Rmd <==
---
title: "4a_lms"
author: "Leslie Roberson"
date: "2/26/2020"
output: html_document
---

## Summary

explore correlations between different variables for the countries:

Does a higher proportion of th species CATCH correlate with a higher percentage of species level records?

multiple linear regression:

vol th spp legal catch ~ vol IUU catch
                  + vol all other spp level import records (not threatened and unknown stats)
                  + vol nonspecific records
                  + region
                  + country governance score (WGI)
                  + GDP per capita (current usd) - better indication of wealth than GDP
                  + population
                  
```{r setup, include=FALSE}

library(tidyverse)

```

## catch

```{r}

## load data
catch <- read_csv("_data/lms_catch_dat.csv")
names(catch)

## LINEAR REGRESSIONS
lm_0 <- lm(spp_th ~ tot_vol, 
                 data = catch)
summary(lm_0)

lm_1 <- lm(spp_th ~ tot_vol + 
             aggregated, 
                 data = catch)
summary(lm_1)

lm_2 <- lm(spp_th ~ tot_vol + 
             aggregated + 
             pop, 
                 data = catch)
summary(lm_2)

lm_3 <- lm(spp_th ~ tot_vol + 
             aggregated + 
             score_wgi_scale, 
                 data = catch)
summary(lm_3)

lm_4 <- lm(spp_th ~ tot_vol + 
             aggregated + 
             gdp_usd, 
                 data = catch)
summary(lm_4)

lm_5 <- lm(spp_th ~ tot_vol + 
             aggregated + 
             gdp_percap_usd, 
                 data = catch)
summary(lm_5)

lm_6 <- lm(spp_th ~ spp_not_th + 
             spp_unknown_status + 
             aggregated + 
             gdp_percap_usd, 
                 data = catch)
summary(lm_6)

lm_7 <- lm(spp_th ~ ratio,
                 data = catch)
summary(lm_7)

lm_8 <- lm(spp_th ~ tot_vol + 
             ratio, 
                 data = catch)
summary(lm_8)

lm_9 <- lm(spp_th ~ tot_vol + 
             gdp_percap_usd, 
                 data = catch)
summary(lm_9) 

## ANOVA best model fit
catch_best <- catch %>%
  filter(!is.na(pop)) %>%
  filter(!is.na(gdp_percap_usd)) 

lm_best <- lm(spp_th ~ tot_vol + 
                aggregated +
                gdp_percap_usd, 
              data = catch_lms_dat_sub_best)
plot(lm_best)

out1 <- summary(lm_best) 
out2 <- as.data.frame(anova(lm_best))
out3 <- as.data.frame(coefficients(lm_best)) 

out4 <- as.data.frame(confint(lm_best, level=0.95)) 
fitted(lm_best) 
vcov(lm_best) 
influence(lm_best) 
  
```

## imports

```{r}

imports <- read_csv("_data/lms_imports_dat.csv")
names(imports)

## LINEAR REGRESSIONS
lm_0 <- lm(spp_th ~ tot_vol, 
                 data = imports)
summary(lm_0)

lm_1 <- lm(spp_th ~ tot_vol + 
             aggregated, 
                 data = imports)
summary(lm_1)

lm_2 <- lm(spp_th ~ tot_vol + 
             aggregated + 
             pop, 
                 data = imports)
summary(lm_2)

lm_3 <- lm(spp_th ~ tot_vol + 
             aggregated + 
             score_wgi_scale, 
                 data = imports)
summary(lm_3)

lm_4 <- lm(spp_th ~ tot_vol + 
             aggregated + 
             gdp_usd, 
                 data = imports)
summary(lm_4)

lm_5 <- lm(spp_th ~ tot_vol + 
             aggregated + 
             gdp_percap_usd, 
                 data = imports)
summary(lm_5)

lm_6 <- lm(spp_th ~ spp_not_th + 
             spp_unknown_status + 
             aggregated + 
             gdp_percap_usd, 
                 data = imports)
summary(lm_6)

lm_7 <- lm(spp_th ~ ratio, 
                 data = imports)
summary(lm_7)

lm_8 <- lm(spp_th ~ tot_vol + 
             ratio, 
                 data = imports)
summary(lm_8) 

## ANOVA
lm_best <- lm(spp_th ~ tot_vol + 
                aggregated,
                 data = imports)
plot(lm_best)
out1 <- summary(lm_best) 
out2 <- as.data.frame(anova(lm_best))
out3 <- as.data.frame(coefficients(lm_best)) 
out4 <- as.data.frame(confint(lm_best, level=0.95))
fitted(lm_best) 
vcov(lm_best) 
influence(lm_best)

```
